+++ src/components/header/header.js
@@ -26,19 +26,20 @@ class Header extends React.Component {
                 {this.props.t('common.title')}
             </Widget.NavbarItem>
 
-            <div className="col-lg-8">
+            <div className="col-lg-7">
             </div>
 
-
-            <Widget.NavbarItem type={Widget.NAVBAR_ITEM.TEXT}>
-                <div className="input-group-sm input-group-prepend">
-                    <label className="col-form-label-sm text-white px-2">{ this.props.t('common.labels.language') }</label>
-                    <Widget.Dropdown onSelect={ this.handleChange }>
-                        <option value="de">{ this.props.t('common.languages.de') }</option>
-                        <option value="en">{ this.props.t('common.languages.en') }</option>
-                    </Widget.Dropdown>
-                </div>
-            </Widget.NavbarItem>
+            <div className="col-lg-1">
+                <Widget.NavbarItem type={Widget.NAVBAR_ITEM.TEXT}>
+                    <div className="input-group-sm input-group-prepend">
+                        <label className="col-form-label-sm text-white px-2">{ this.props.t('common.labels.language') }</label>
+                        <Widget.Dropdown onSelect={ this.handleChange }>
+                            <option value="de">{ this.props.t('common.languages.de') }</option>
+                            <option value="en">{ this.props.t('common.languages.en') }</option>
+                        </Widget.Dropdown>
+                    </div>
+                </Widget.NavbarItem>
+            </div>
         </Widget.Navbar>;
     }
 }