+++ src/pages/kitchensink/Kitchensink.js
@@ -1,0 +1,27 @@
+import React from "react";
+import DropDownExample from "./DropDownExample";
+import {withTranslation} from "react-i18next";
+import ButtonExample from "./ButtonExample";
+import ExpandableSectionExample from "./ExpandableSectionExample";
+import TwoWayDatabingExample from "./TwoWayDatabingExample";
+import ListExample from "./ListExample";
+
+class Kitchensink extends React.Component {
+
+    test() {
+        alert("Test");
+    }
+
+    render() {
+        return <div>
+            <h1> { this.props.t('app.kitchensink.title') } </h1>
+            <ButtonExample handler={ this.test }/>
+            <DropDownExample/>
+            <ExpandableSectionExample handler={ this.test }/>
+            <TwoWayDatabingExample/>
+            <ListExample/>
+        </div>
+    }
+}
+
+export default withTranslation()(Kitchensink);