+++ src/pages/kitchensink/Mockdata.js
@@ -1,0 +1,28 @@
+
+export const WorldData = [
+    {
+        "id": 1,
+        "name": "Mercury",
+        "diameter": 4879,
+        "greeting": "Pipl Hi"
+    },
+    {
+        "id": 2,
+        "name": "Venus",
+        "diameter": 12103,
+        "greeting": "Shum na!"
+    },
+    {
+        "id": 3,
+        "name": "Earth",
+        "diameter": 12756,
+        "greeting": "Hello world!"
+    },
+    {
+        "id": 4,
+        "name": "Mars",
+        "diameter": 6792,
+        "greeting": "Tach pach!"
+    }
+    ];
+