+++ src/shared/components/index.js
@@ -3,15 +3,18 @@ import WNavbar, {NB} from "./wdg.navbar";
 import {APPEARANCE, NB_ITEM_TYPE, SIZE} from "./constants";
 import {WDropdown} from "./wdg.dropdown";
 import {WExpandableBody, WExpandableSection} from "./wdg.expandable-tab";
+import {WList} from "./wdg.list";
 
 export const Widget = {
 
     Navbar: WNavbar,
     NavbarItem: NB.item,
+
     Button: WButton,
     Dropdown: WDropdown,
     ExpandableSection: WExpandableSection,
     ExpandableTabBody: WExpandableBody,
+    List: WList,
 
     BUTTON_TYPE: APPEARANCE,
     SIZE: SIZE,